+++ fronts-client/src/components/modals/OptionsModal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ButtonDefault from 'components/inputs/ButtonDefault';
-import { OptionsModalProps } from '../types/Modals';
+import { OptionsModalProps } from '../../types/Modals';
 import type { State } from 'types/State';
 import {
   selectOptionsModalIsOpen,
